+++ src/database/domains.js
@@ -54,6 +54,7 @@ export const update = async (id, data) => {
       },
       {
         returnDocument: 'after',
+        runValidators: true,
       },
     ),
   )