+++ src/models/Action.js
@@ -15,13 +15,17 @@ const schema = new mongoose.Schema({
   },
   key: {
     type: String,
+    maxlength: 500,
   },
   value: {
     type: Number,
-    required: true,
+    required() {
+      return this.isNew === true
+    },
   },
   details: {
     type: String,
+    maxlength: 2000,
   },
   created: {
     type: Date,