+++ src/models/Event.js
@@ -21,6 +21,7 @@ const schema = new mongoose.Schema({
   title: {
     type: String,
     required: true,
+    maxlength: 500,
   },
   type: {
     type: String,